+++ utilities/.eslintrc.cjs
@@ -46,7 +46,7 @@ module.exports = {
                     'body': 1,
                     'parameters': 2,
                 },
-                'MemberExpression': 2,
+                'MemberExpression': 1,
                 'ObjectExpression': 1,
                 'SwitchCase': 1,
                 'ignoredNodes': [